Checnged a lot of code
Changed a lot of code so there a less request to a website. Multiple options for websites including subdomain, IP address in URL are supported.
New functions added as well such as showing the IP address, take a screenshot creating a file in a collected-data directory which contains all data which is collected.

In the near future the -l (list in a file) options we be developed. With this option several websites could be checked.